make compiler happy

+++ BadvpnTun2socks/tun2socks-iOS/misc/ipaddr6.h
@@ -214,7 +214,7 @@ int ipaddr6_parse_ipv6_prefix (MemRef str, int *out_num)
         return 0;
     }
     
-    *out_num = d;
+    *out_num = (int) d;
     return 1;
 }
